Autonomous Electric Rickshaw
Application: Autonomous Electric Rickshaw for First and Last Miles

This document is a collective effort of 31st batch students of Madras Institute of Technology,
Chromepet, Anna University, Chennai, Tamil Nadu, India.

Mission: develop innovation and entrepreneurship in student community

Innovation: Needs have always been there; most of the needs are satisfied; some of the needs
are not satisfied yet. Unsatisfied needs are the seeds for innovation.